// src/app/models/booking.model.ts
export interface Booking {
  id?: string;
  contactName: string;
  package: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  country: string;
  zipcode: string;
  phone1: string;
  phone2: string;
  fax: string;
  email1: string;
  email2: string;
  notes: string;
}